+++ src/Anamorphosis/AnamDiscreteIR.cpp
@@ -96,7 +96,7 @@ int AnamDiscreteIR::fit(const VectorDouble& tab, int verbose)
   /* Initializations */
 
   error = 1;
-  int nech = tab.size();
+  int nech = static_cast<int> (tab.size());
   int nclass = getNClass();
   int ncut = getNCut();
   residuals = T = Q = (double *) NULL;
